biggest change is that the formerly default view is now only accessible through sub.enketo.formhub.org/classic. The new default view is an almost empty front page.
