+++ Code_Igniter/application/controllers/html5validate.php
@@ -32,13 +32,6 @@ class Html5validate extends CI_Controller {
 		            'content'=>$content
 		        );
 
-		//url-ify the data for the POST
-		//foreach($fields as $key=>$value) 
-		//{ 
-		//	$fields_string .= $key.'='.$value.'&'; 
-		//}
-		//rtrim($fields_string,'&');
-
 		//open connection
 		$ch = curl_init();
 
@@ -55,36 +48,6 @@ class Html5validate extends CI_Controller {
 
 		echo $result;
 
-		//$this->load->helper('subdomain');
-//		$this->load->helper('url');
-//		$subdomain = get_subdomain(); //from subdomain helper
-//		
-//		//$this->load->model('Form_model','',TRUE);
-//				
-//		if (isset($subdomain))
-//		{
-//			show_404();
-//		}
-//		else 
-//		{
-//			$data = array('offline'=>FALSE, 'title_component'=>'launch');
-//			if (ENVIRONMENT === 'production')
-//			{
-//				$data['scripts'] = array(
-//					base_url('js-min/launch-all-min.js')
-//				);
-//			}
-//			else
-//			{
-//				$data['scripts'] = array(
-//					base_url('js-source/__common.js'),
-//					base_url('js-source/__form.js'),
-//					base_url('js-source/__launch.js'),
-//					base_url('js-source/__debug.js')
-//				);
-//			}
-//			$this->load->view('launch_view', $data);
-//		}
 	}
 }
 ?>
